DDF-3343 - Splits Metacard model up
 - Splits Metacard model up to be more manageable / refactorable in the future.
